requirements.txt
# Core
pyyaml>=6.0
pydantic>=2.0
python-dotenv>=1.0

# LLM clients
google-genai>=1.0
anthropic>=0.40

# Database (Postgres + pgvector)
asyncpg>=0.30
psycopg[binary]>=3.1
pgvector>=0.3

# Retrieval
flashrank>=0.2

# Dashboard
aiohttp>=3.9

# Utilities
numpy>=1.26
python-dateutil>=2.8
